실행환경 맞춤 업데이트: 의존성/모듈/웹 UI 정리
- requirements.txt를 실제 동작 환경(conda base) 설치 버전 기준으로 갱신
- modules 분석 코드 및 web UI(interview/index/style 등) 개선
- 빈 null 파일 제거

+++ modules/evaluation/audio_metrics.py
@@ -29,9 +29,7 @@ def _read_wav_mono_int16(path: str) -> Tuple[np.ndarray, int]:
 
     audio = np.frombuffer(raw, dtype=np.int16)
     if nch > 1:
-        # float32 로 평균내 채널 합산 시 int16 오버플로/정밀도 손실을 피한다.
-        # (RMS 계산도 어차피 float 로 하므로 int16 으로 되돌리지 않는다.)
-        audio = audio.reshape(-1, nch).astype(np.float32).mean(axis=1)
+        audio = audio.reshape(-1, nch).mean(axis=1).astype(np.int16)
 
     return audio, sr
